+++ CHANGELOG.md
@@ -14,8 +14,11 @@ and this project adheres to
     colors when this variable is set with non-empty value (#45).
   - Support optional logger component used as prefix of every log entries.
 - settings: Support parameters deprecation (#48).
-- auth: Introduce `JWTBaseManager` for more generic and versatile JWT encode and
-  decode management with custom claimset.
+- auth:
+  - Introduce `JWTBaseManager` for more generic and versatile JWT encode and
+    decode management with custom claimset.
+  - Introduce `jwt_validate_expiration()` function to load and validate
+    expiration of JWT without its signature.
 
 ### Changed
 - log: Disable debug by default with `DaemonFormatter` similarly to